Разделила main.js на модули
