+++ js/const.js
@@ -1,0 +1,57 @@
+const COUNT = 10;
+
+const TITLE = [
+  'Уютное гнездышко для молодоженов',
+  'Маленькая квартирка рядом с парком',
+  'Небольшая лавочка в парке',
+  'Императорский дворец в центре Токио',
+  'Милейший чердачок',
+  'Наркоманский притон',
+  'Чёткая хата',
+  'Стандартная квартира в центре',
+  'Тихая квартирка недалеко от метро',
+  'Милое гнездышко для фанатов Анимэ',
+];
+
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+
+const TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESCRIPTION = [
+  'Великолепный таун-хауc в центре Токио. Подходит как туристам, так и бизнесменам. Дом полностью укомплектован и имеет свежий ремонт.',
+  'Маленькая чистая квратира на краю парка. Без интернета, регистрации и СМС.',
+  'Великолепная лавочка прямо в центре парка. Подходит для всех кто любит спать на свежем воздухе.',
+  'Замечательный дворец в старинном центре города. Только для тех кто может себе позволить дворец. Лакеев и прочих жокеев просим не беспокоить.',
+  'Маленькая квартирка на чердаке. Для самых нетребовательных.',
+  'У нас есть всё! Шприцы, интернет, кофе. Для всех кто знает толк в отдыхе. Полицию просим не беспокоить.',
+  'У нас тут все ништяк. Ларек за углом. Шава 24 часа. Приезжайте! Интернетов нет!',
+  'Тут красиво, светло и уютно. Есть где разместиться компании из 5 человек. Кофе и печеньки бесплатно.',
+  'Квартира на первом этаже. Соседи тихие. Для всех, кто терпеть не может шума и суеты.',
+  'Хейтеров просьба не беспокоить.',
+];
+
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+export {COUNT, TITLE, TYPE, TIME, FEATURES, DESCRIPTION, PHOTOS};
